KH_Gender_Bias1/api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pandas
openai
textdistance
nltk
openpyxl
setuptools==68.2.2
glibc==0.6.1
tiktoken
python-dotenv